refactor: domain repository 사용하도록 수정
